requirements.txt
google-generativeai
pandas
openpyxl
xlrd
logging